refactor(control-plane): split handlers/ui/dashboard.go into focused files
Splits the 1493 LOC dashboard.go (29 functions, 20+ types, 4 mixed
concerns) into four cohesive files within the same handlers/ui package:

  dashboard.go           (348 LOC) - DashboardHandler + 2 public handlers
                                      + request-parsing helpers + base
                                      response types
  dashboard_cache.go     (146 LOC) - DashboardCache, EnhancedDashboardCache,
                                      TimeRangePreset, cache key/TTL utils
  dashboard_format.go    (177 LOC) - response-shape type declarations
  dashboard_aggregate.go (839 LOC) - build*/get* aggregators and helpers

Pure file moves within package ui: no renames, no behavior change, no
exported-API change. All imports and callers (server.go route
registration, same-package tests) resolve unchanged.

Prep work for the WP7 test-coverage push — smaller files give tests a
narrower surface to target.

Closes #417
